+++ include/Objects/Purchase.hpp
@@ -11,11 +11,11 @@ class Purchase
     std::vector<Fragrance> fragrances;
     PurchaseStatus status;
     int userID;
-    float finalPrice;
+    long finalPrice;
 
 public:
     // will use move for fragrances
-    Purchase(size_t id, std::vector<Fragrance> fragrances, PurchaseStatus status, size_t userID, float finalPrice = 0);
+    Purchase(size_t id, std::vector<Fragrance> fragrances, PurchaseStatus status, size_t userID, long finalPrice = 0);
 
     void setStatus(PurchaseStatus newStatus);
 
@@ -23,7 +23,7 @@ public:
     std::vector<Fragrance> getFragrances() const;
     PurchaseStatus getStatus() const;
     size_t getUserID() const;
-    float getFinalPrice() const;
+    long getFinalPrice() const;
 
     void show(const std::string &userName) const;
     void serialize(std::ostream &os) const;
